refactor: extract useSidebar
