+++ src/components/App/use-sidebar.js
@@ -1,0 +1,16 @@
+import { useState } from 'react'
+
+const useSidebar = () => {
+    const [visible, setVisible] = useState(true)
+    return {
+        visible,
+        show: () => {
+            setVisible(true)
+        },
+        hide: () => {
+            setVisible(false)
+        },
+    }
+}
+
+export default useSidebar
